Properly handle complex queries in ModuleSpace
Break complex query on a single ones. Apply each query to self and merge
results.
