+++ lib/src/space/module.rs
@@ -25,6 +25,10 @@ impl ModuleSpace {
     }
 
     pub fn query(&self, query: &Atom) -> BindingsSet {
+        complex_query(query, |query| self.single_query(query))
+    }
+ 
+    fn single_query(&self, query: &Atom) -> BindingsSet {
         log::debug!("ModuleSpace::query: {} {}", self, query);
         let mut results = self.main.query(query);
         for dep in &self.deps {
@@ -88,6 +92,28 @@ impl SpaceMut for ModuleSpace {
     }
     fn as_space<'a>(&self) -> &(dyn Space + 'a) {
         self
+    }
+}
+
+#[cfg(test)]
+mod test {
+    use crate::*;
+    use crate::space::grounding::*;
+    use super::*;
+
+    #[test]
+    fn complex_query_two_subspaces() {
+        let mut a = GroundingSpace::new();
+        a.add(expr!("a" "b"));
+        let mut b = GroundingSpace::new();
+        b.add(expr!("b" "c"));
+
+        let mut main = ModuleSpace::new(GroundingSpace::new());
+        main.add_dep(DynSpace::new(ModuleSpace::new(a)));
+        main.add_dep(DynSpace::new(ModuleSpace::new(b)));
+
+        assert_eq_no_order!(main.query(&expr!("," (a "b") ("b" c))), vec![bind!{ a: sym!("a"), c: sym!("c") }]);
+        assert_eq_no_order!(main.query(&expr!("," ("a" b) (b "c"))), vec![bind!{ b: sym!("b") }]);
     }
 }
